refactor: add player strategy interface
