+++ src/main/java/org/bitsquad/warzone/player/PlayerStrategy.java
@@ -1,0 +1,5 @@
+package org.bitsquad.warzone.player;
+
+public interface PlayerStrategy {
+    void issueOrder();
+}
